freelance_platform/requirements.txt
flask==2.3.3
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
flask-jwt-extended==4.6.0
flask-cors==4.0.0
werkzeug==2.3.7
python-dotenv==1.0.0
marshmallow==3.20.1
requests==2.31.0
pytest==7.4.3
gunicorn==21.2.0
sqlalchemy==2.0.25
